IOTbackend/requirements.txt
django
pyjwt
